Fix UI border-radius

+++ client/src/components/View/Content/components/ContextMenu/ContextMenu.js
@@ -5,23 +5,26 @@ function ContextMenu({ visible, dataEvent, cardData, setContent, handleCloseCont
 	if (dataEvent) {
 		cntxtMenu.current = {
 			left: `${dataEvent.clientX}px`,
-			top: `${dataEvent.clientY}px`,
+			top: `${dataEvent.clientY - 10}px`,
 		};
 	}
 	function deleteCard(event) {
-		fetch(`${process.env.REACT_APP_DESTINATION_REQUEST}/${cardData._id}/delete`, {
-			method: "DELETE",
-		}).then((res) => {
-			if (res.status === 200) {
-				alert("Delete successfully!");
-				setContent((pre) => {
-					return pre ? false : true;
-				});
-				handleCloseContextMenu(event);
-			} else {
-				alert("Delete failure!");
-			}
-		});
+		const isDeleteCard = window.confirm(`Accept to delete this card?\nCard title: ${cardData.title}`);
+		if (isDeleteCard) {
+			fetch(`${process.env.REACT_APP_DESTINATION_REQUEST}/${cardData._id}/delete`, {
+				method: "DELETE",
+			}).then((res) => {
+				if (res.status === 200) {
+					alert("Delete successfully!");
+					setContent((pre) => {
+						return pre ? false : true;
+					});
+					handleCloseContextMenu(event);
+				} else {
+					alert("Delete failure!");
+				}
+			});
+		}
 	}
 	function handleUpdateCardContent(event) {
 		const { _id, status, title, description } = cardData;
@@ -35,7 +38,7 @@ function ContextMenu({ visible, dataEvent, cardData, setContent, handleCloseCont
 		});
 	}
 	return visible ? (
-		<div style={cntxtMenu.current} className="absolute flex flex-col bg-white p-3 rounded drop-shadow-md translate-x-[-100%] translate-y-[-100%]">
+		<div style={cntxtMenu.current} className="absolute flex flex-col bg-white p-3 xl:rounded-sm rounded drop-shadow-md translate-x-[-100%] translate-y-[-100%]">
 			<div onClick={handleUpdateCardContent} className="py-1 px-2 text-sm hover:bg-slate-100 cursor-pointer">
 				Update
 			</div>
